temp_test.py
import os
from dotenv import load_dotenv
from google import genai

load_dotenv()
key = os.getenv('GEMINI_API_KEY')

try:
    client = genai.Client(api_key=key)
    for m in client.models.list():
        print(m.name)
except Exception as e:
    import traceback
    traceback.print_exc()
