requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
faster-whisper>=1.0.0
google-genai
python-dotenv>=1.0.0
